+++ build-asan/examples/CMakeFiles/system_example.dir/cmake_clean.cmake
@@ -1,0 +1,11 @@
+file(REMOVE_RECURSE
+  "CMakeFiles/system_example.dir/system_example.cpp.o"
+  "CMakeFiles/system_example.dir/system_example.cpp.o.d"
+  "system_example"
+  "system_example.pdb"
+)
+
+# Per-language clean rules from dependency scanning.
+foreach(lang CXX)
+  include(CMakeFiles/system_example.dir/cmake_clean_${lang}.cmake OPTIONAL)
+endforeach()