+++ build-asan/examples/CMakeFiles/telemetry_bus_example.dir/cmake_clean.cmake
@@ -1,0 +1,11 @@
+file(REMOVE_RECURSE
+  "CMakeFiles/telemetry_bus_example.dir/telemetry_bus_example.cpp.o"
+  "CMakeFiles/telemetry_bus_example.dir/telemetry_bus_example.cpp.o.d"
+  "telemetry_bus_example"
+  "telemetry_bus_example.pdb"
+)
+
+# Per-language clean rules from dependency scanning.
+foreach(lang CXX)
+  include(CMakeFiles/telemetry_bus_example.dir/cmake_clean_${lang}.cmake OPTIONAL)
+endforeach()