+++ build-asan/examples/CMakeFiles/timing_controller_example.dir/cmake_clean.cmake
@@ -1,0 +1,11 @@
+file(REMOVE_RECURSE
+  "CMakeFiles/timing_controller_example.dir/timing_controller_example.cpp.o"
+  "CMakeFiles/timing_controller_example.dir/timing_controller_example.cpp.o.d"
+  "timing_controller_example"
+  "timing_controller_example.pdb"
+)
+
+# Per-language clean rules from dependency scanning.
+foreach(lang CXX)
+  include(CMakeFiles/timing_controller_example.dir/cmake_clean_${lang}.cmake OPTIONAL)
+endforeach()