+++ build-asan/examples/CTestTestfile.cmake
@@ -1,0 +1,6 @@
+# CMake generated Testfile for 
+# Source directory: /home/ag7/Documents/AxonVex/examples
+# Build directory: /home/ag7/Documents/AxonVex/build-asan/examples
+# 
+# This file includes the relevant testing commands required for 
+# testing this directory and lists subdirectories to be tested as well.