+++ build-tsan/examples/CMakeFiles/memory_pool_ports_simple_example.dir/cmake_clean.cmake
@@ -1,0 +1,11 @@
+file(REMOVE_RECURSE
+  "CMakeFiles/memory_pool_ports_simple_example.dir/memory_pool_ports_simple_example.cpp.o"
+  "CMakeFiles/memory_pool_ports_simple_example.dir/memory_pool_ports_simple_example.cpp.o.d"
+  "memory_pool_ports_simple_example"
+  "memory_pool_ports_simple_example.pdb"
+)
+
+# Per-language clean rules from dependency scanning.
+foreach(lang CXX)
+  include(CMakeFiles/memory_pool_ports_simple_example.dir/cmake_clean_${lang}.cmake OPTIONAL)
+endforeach()