+++ build-tsan/examples/CMakeFiles/simple_callback_demo.dir/cmake_clean.cmake
@@ -1,0 +1,11 @@
+file(REMOVE_RECURSE
+  "CMakeFiles/simple_callback_demo.dir/simple_callback_demo.cpp.o"
+  "CMakeFiles/simple_callback_demo.dir/simple_callback_demo.cpp.o.d"
+  "simple_callback_demo"
+  "simple_callback_demo.pdb"
+)
+
+# Per-language clean rules from dependency scanning.
+foreach(lang CXX)
+  include(CMakeFiles/simple_callback_demo.dir/cmake_clean_${lang}.cmake OPTIONAL)
+endforeach()