+++ build-tsan/examples/CMakeFiles/threadSafeQueue_example.dir/cmake_clean.cmake
@@ -1,0 +1,11 @@
+file(REMOVE_RECURSE
+  "CMakeFiles/threadSafeQueue_example.dir/threadSafeQueueExample.cpp.o"
+  "CMakeFiles/threadSafeQueue_example.dir/threadSafeQueueExample.cpp.o.d"
+  "threadSafeQueue_example"
+  "threadSafeQueue_example.pdb"
+)
+
+# Per-language clean rules from dependency scanning.
+foreach(lang CXX)
+  include(CMakeFiles/threadSafeQueue_example.dir/cmake_clean_${lang}.cmake OPTIONAL)
+endforeach()